+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Unreleased
 ----------
+- Introduced `--tag` option to various sub-commands to "tag" snapshots
+  of the same subvolume for different backup destinations
 - Fixed `--help` & `--version` options to no longer report errors